+++ src/Compiler/LatteToPhpCompiler.php
@@ -6,6 +6,7 @@ namespace Efabrica\PHPStanLatte\Compiler;
 
 use Composer\InstalledVersions;
 use Efabrica\PHPStanLatte\Compiler\Compiler\CompilerInterface;
+use Efabrica\PHPStanLatte\Exception\ParseException;
 use Efabrica\PHPStanLatte\Template\Template;
 use InvalidArgumentException;
 use Latte\CompileException;
@@ -43,6 +44,7 @@ final class LatteToPhpCompiler
 
     /**
      * @throws CompileException
+     * @throws ParseException
      */
     public function compileFile(Template $template, string $context = ''): string
     {